config: Clean up and document parse.go
Fix improper handling of closing brace.
Now parser can handle the following example correctly:
a {
 b }

And also this:
a { b }
c
